backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
# PostgreSQL driver (psycopg3); use postgresql+psycopg:// in DATABASE_URL
psycopg[binary]>=3.2.0
alembic>=1.14.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
feedparser>=6.0.11
pyyaml>=6.0.0
beautifulsoup4>=4.12.0
